+++ grace_addon/files/general/www/public/company_licenses.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'auth.php';
 $pageTitle = 'GRACe - Company Licenses';
 $useJquery = true;
 require 'header.php';