+++ grace_addon/files/general/www/public/current_plants.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'auth.php';
 $pageTitle = 'GRACe - Current Plants';
 require 'header.php';
 ?>