+++ grace_addon/files/general/www/public/get_rooms.php
@@ -1,4 +1,3 @@
-<?php require_once 'auth.php'; ?>
 <?php
 require_once 'config.php';