+++ grace_addon/files/general/www/public/record_dry_weight.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'auth.php';
 $pageTitle = 'GRACe - Record Flower Transaction';
 require 'header.php';
 ?>