+++ grace_addon/files/general/www/public/reporting.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'auth.php';
 $pageTitle = 'GRACe - Reporting';
 require 'header.php';
 ?>