+++ grace_addon/files/general/www/public/show_database.php
@@ -1,4 +1,3 @@
-<?php require_once 'auth.php'; ?>
 <?php
 ini_set('display_errors', 1);
 ini_set('display_startup_errors', 1);